Add command aliases to improve consistency with the Logo books
Added various command aliases to improve consistency with the Logo
books, for example "Apple Logo" by Abelson, "Logo Works: Lessons in
Logo" by Cory and Walker, and "Teaching with Logo" by Watt and Watt.
For example, turnleft is now aliased as left and lt.
Add default parameter of 90 to turnleft and turnright